+++ labpack/datasets/__init__.py
@@ -1,4 +1,4 @@
-''' a folder to contain required data tables and package of methods for accessing them '''
+''' a folder to contain required data tables and packages of methods for accessing them '''
 __author__ = 'rcj1492'
 __created__ = '2017.10'
 __license__ = 'MIT'